pyproject.toml
[project]
name = "remlight"
version = "0.1.0"
description = "Minimal declarative agent framework with PostgreSQL memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10",
    "pydantic-ai>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "asyncpg>=0.30",
    "fastmcp>=0.5",
    "click>=8.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
]

[project.scripts]
rem = "remlight.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["remlight"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--asyncio-mode=auto",
]
markers = [
    "asyncio: marks tests as async",
    "unit: marks tests as unit tests (no external dependencies)",
    "integration: marks tests as integration tests (require database)",
    "llm: marks tests that make LLM API calls (expensive, skip by default with -m 'not llm')",
    "slow: marks tests as slow running",
]
